+++ src/fri/mod.rs
@@ -1,57 +1,8 @@
-use crate::field::Field128;
 use crate::merkle_tree::{Merkle, MerkleInclusionPath, MerkleInclusionPathError};
 use crate::ntt::{NttField, Polynomial};
+use crate::transcript::{HashableField, Transcript};
 use crate::{field::Field, merkle_tree::HashDigest};
 use serde::{Deserialize, Serialize};
-
-use sha2::{Digest, Sha256};
-
-pub struct Transcript {
-    state: Sha256,
-}
-
-impl Default for Transcript {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Transcript {
-    pub fn new() -> Self {
-        Self {
-            state: Sha256::new(),
-        }
-    }
-
-    pub fn append_message(&mut self, label: &[u8], message: &[u8]) {
-        self.state.update(label);
-        self.state.update(message);
-    }
-
-    pub fn random(&self) -> [u8; 32] {
-        let cloned_state = self.state.clone();
-        let result = cloned_state.finalize();
-        let mut random_bytes = [0u8; 32];
-        random_bytes.copy_from_slice(&result[..32]);
-        random_bytes
-    }
-}
-
-pub trait HashableField: Field + AsRef<[u8]> {
-    fn to_bytes(&self) -> &[u8];
-    fn from_digest(digest: &[u8; 32]) -> Self;
-}
-
-impl HashableField for Field128 {
-    fn to_bytes(&self) -> &[u8] {
-        self.as_ref()
-    }
-
-    fn from_digest(digest: &[u8; 32]) -> Self {
-        let x = u128::from_le_bytes(digest[0..16].try_into().unwrap());
-        Self::from(x)
-    }
-}
 
 pub struct ProverData<F> {
     pub commitments: Vec<Merkle<F>>,
@@ -229,7 +180,7 @@ impl<F: HashableField> ProverData<F> {
             paths.push((path, conjugate_path));
 
             current_n /= 2;
-            current_index = current_index % current_n;
+            current_index %= current_n;
             current_conjugate = (current_index + current_n / 2) % current_n;
         }
 
@@ -297,7 +248,7 @@ impl<F: HashableField + NttField> QueryProof<F> {
 
             current_gen *= current_gen;
             current_n /= 2;
-            current_index = current_index % current_n;
+            current_index %= current_n;
             current_conjugate = (current_index + current_n / 2) % current_n;
         }
 
@@ -405,6 +356,7 @@ impl<F: HashableField + NttField> FriProof<F> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::field::Field128;
     use crate::ntt::NttField;
     use bincode; // Ensure bincode is imported
