+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod field;
 pub mod fri;
 pub mod merkle_tree;
 pub mod ntt;
+pub mod transcript;
